+++ tests/main.py
@@ -20,7 +20,7 @@ import os
 import time
 import argparse
 import torch
-
+import mcr_dl
 from mcr_dl.constants import TORCH_DISTRIBUTED_DEFAULT_PORT
 from common import set_accelerator_visible
 from mcr_dl.cuda_accelerator import get_accelerator
@@ -32,38 +32,8 @@ parser.add_argument("--backend", choices=['mpi', 'nccl'], help = "Backend")
 parser.add_argument("--dist", choices=['mcr_dl', 'torch'], help = "torch.distributed or mcr-dl for distributed")
 args = parser.parse_args()
 
-def init_torch_distributed(backend):
-    global dist
-    import torch.distributed as dist
-    if backend == 'nccl':
-        mpi_discovery()
-    elif backend == 'mpi':
-        set_mpi_dist_environemnt()
-    dist.init_process_group(backend)
-    local_rank = int(os.environ['LOCAL_RANK'])
-    get_accelerator().set_device(local_rank)
-
-
-def init_mcr_dl_comm(backend):
-    global dist
-    import mcr_dl
-    import mcr_dl as dist
-    mcr_dl.init_distributed(dist_backend=backend, use_mcr_dl=True)
-    local_rank = int(os.environ['LOCAL_RANK'])
-    #get_accelerator().set_device(local_rank)
-
-
-def init_processes(args_dist, args_backend):
-    print(f'Comm : {args_dist}  Backend : {args_backend}')
-    if args_dist == 'mcr_dl':
-        init_mcr_dl_comm(args_backend)
-    elif args_dist == 'torch':
-        init_torch_distributed(args_backend)
-    else:
-        print(f"distributed framework {args_dist} not supported")
-        exit(0)
-
 def all_reduce():
+    dist = mcr_dl.get_distributed_engine()
     x = torch.ones(1, 3).to(get_accelerator().device_name()) * (dist.get_rank() + 1)
     sum_of_ranks = (dist.get_world_size() * (dist.get_world_size() + 1)) // 2
     result = torch.ones(1, 3).to(get_accelerator().device_name()) * sum_of_ranks
@@ -71,6 +41,7 @@ def all_reduce():
     assert torch.all(x == result)
 
 def all_reduce_benchmark():
+    dist = mcr_dl.get_distributed_engine()
     start_events = [torch.cuda.Event(enable_timing=True) for _ in range(2, 30)]
     end_events = [torch.cuda.Event(enable_timing=True) for _ in range(2, 30)]
     rank = dist.get_rank()
@@ -101,7 +72,7 @@ def all_reduce_benchmark():
 
 if __name__ == "__main__":
     set_accelerator_visible()
-    init_processes(args_dist=args.dist, args_backend=args.backend)
+    mcr_dl.init_processes(dist_engine = args.dist, dist_backend = args.backend)
     # all_reduce()
     all_reduce_benchmark()
